# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Streambox API
  description: API REST para la gestión de películas, géneros y usuarios de Streambox.
  version: '1.0'
servers:
- url: http://localhost:8080
  description: Backend local (perfil dev)
paths:
...
  /api/users:
    get:
      tags:
      - user-controller
      summary: Obtiene todos los usuarios
      description: Devuelve la lista de usuarios registrados. Este endpoint requiere permisos de administrador.
      operationId: getUsers
      responses:
        '200':
          description: Usuarios obtenidos correctamente
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          description: El usuario no está autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: El usuario no tiene permisos de administrador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
      - bearerAuth: []
    post:
      tags:
      - user-controller
      summary: Registra un nuevo usuario
      description: Crea una nueva cuenta de usuario en Streambox. No requiere autenticación.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '201':
          description: Usuario creado correctamente
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 'Los datos proporcionados no son válidos (código VALIDATION_ERROR, un mensaje por campo en validationErrors).
            La contraseña debe tener entre 12 y 64 caracteres, ocupar como máximo 72 bytes en UTF-8, no ser común ni trivial
            y no contener el nombre de usuario ni la parte local del email. Si incumple varias reglas solo se informa de una,
            por este orden: longitud, bytes, común, datos personales'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: El nombre de usuario o el correo ya están en uso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Demasiados registros desde esta IP; ver cabecera Retry-After
          headers:
            Retry-After:
              description: Segundos que hay que esperar antes de volver a intentarlo (entero, mínimo 1).
              required: true
              schema:
                type: integer
                format: int32
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/users/me:
    get:
      tags:
      - user-controller
      summary: Obtiene el usuario autenticado
      description: Devuelve la información del usuario asociado al token JWT utilizado en la petición.
      operationId: getCurrentUser
      responses:
        '200':
          description: Usuario obtenido correctamente
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: El usuario no está autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
      - bearerAuth: []
    patch:
      tags:
      - user-controller
      summary: Cambia mi nombre de usuario
      description: 'Cambia el nombre de usuario de la cuenta autenticada. El nombre se normaliza (sin espacios ni caracteres
        invisibles en los extremos y con los espacios interiores repetidos reducidos a uno) y debe tener entre 3 y 50 caracteres.
        Si coincide con el actual se responde 200 sin cambios. El correo electrónico, el rol y la contraseña no se pueden
        cambiar aquí: enviarlos da 400.'
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
        required: true
      responses:
        '200':
          description: Nombre cambiado (o sin cambios si era el mismo)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 'Datos no válidos (código VALIDATION_ERROR, un mensaje por campo en validationErrors): nombre ausente
            o de longitud no válida tras normalizarlo, o se ha enviado email, role o password'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: El usuario no está autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Código CSRF_REJECTED: la petición se autentica por la cookie streambox_token y le falta la cabecera
            X-Requested-With: StreamBox (defensa CSRF). Con Authorization: Bearer no se exige'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: El nombre de usuario ya está en uso (código USER_ALREADY_EXISTS)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
      - bearerAuth: []
...
components:
  schemas:
...
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 3
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          description: 'Contraseña de la cuenta: entre 12 y 64 caracteres y como máximo 72 bytes en UTF-8 (las tildes y eñes
            ocupan 2 bytes y los emojis 4). No puede ser una contraseña común o trivial ni contener el nombre de usuario o
            la parte del email anterior a la @. No se exigen mayúsculas, números ni símbolos: se recomienda una frase larga.'
          example: un paseo por el retiro en otoño
          maxLength: 64
          minLength: 12
      required:
      - email
      - password
      - username
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - USER
          - ADMIN
        createdAt:
          type: string
          format: date-time
...
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          description: Nuevo nombre de usuario. Se quitan los espacios (y caracteres invisibles) de los extremos y los interiores
            repetidos se reducen a uno; tras ello debe tener entre 3 y 50 caracteres y no estar en uso por otra cuenta.
          example: cinefila_88
          maxLength: 50
          minLength: 3
      required:
      - username
...
    ErrorResponse:
      type: object
      description: 'Formato común de todas las respuestas de error de la API (4xx y 5xx). El cliente debe decidir por status
        y code, nunca por el texto de message. Los campos que no aplican a un error se omiten del JSON: nunca se envían con
        valor null.'
      properties:
        timestamp:
          type: string
          format: date-time
          description: Fecha y hora del error, en UTC y formato ISO-8601.
          example: '2026-10-04T11:15:30.123Z'
        status:
          type: integer
          format: int32
          description: Código de estado HTTP; coincide con el de la respuesta.
          example: 404
        error:
          type: string
          description: Frase estándar del estado HTTP según la especificación (en inglés).
          example: Not Found
        code:
          type: string
          description: 'Código funcional estable del error: el cliente debe decidir por este campo (y por status), nunca por
            el texto de message. Valores:

            - RESOURCE_NOT_FOUND (404): el recurso o la ruta no existen (una serie sin episodios también da este código a
            los usuarios).

            - MOVIE_NOT_IN_FAVORITES (404): la película no está en la lista del usuario.

            - SERIES_NOT_IN_FAVORITES (404): la serie existe, pero no está en la lista del usuario.

            - CSRF_REJECTED (403): petición no segura (POST/PUT/PATCH/DELETE) autenticada por cookie sin la cabecera X-Requested-With:
            StreamBox. POST /api/auth/refresh y POST /api/auth/logout la exigen siempre, también con Authorization: Bearer.

            - USER_ALREADY_EXISTS (409): el nombre de usuario o el email ya están registrados.

            - MOVIE_ALREADY_IN_FAVORITES (409): la película ya está en la lista del usuario.

            - SERIES_ALREADY_IN_FAVORITES (409): la serie ya está en la lista del usuario.

            - EPISODE_ALREADY_EXISTS (409): la serie ya tiene un episodio con esa temporada y número.

            - GENRE_ALREADY_EXISTS (409): ya existe otro género con ese nombre.

            - GENRE_IN_USE (409): el género no se puede borrar porque alguna película o serie lo tiene asignado.

            - AMBIGUOUS_TITLE (409): varias películas tienen ese título; hay que usar el id.

            - DATA_INTEGRITY_VIOLATION (409): la operación choca con una restricción de los datos.

            - VALIDATION_ERROR (400): datos o parámetros no válidos; el detalle por campo va en validationErrors.

            - CURRENT_PASSWORD_INCORRECT (400): al cambiar la contraseña, la actual no es correcta; el detalle va en validationErrors.currentPassword.
            No es un 401: la sesión sigue siendo válida.

            - MALFORMED_REQUEST (400 y otros 4xx genéricos): la petición no se puede interpretar (JSON mal formado, tipos
            incorrectos).

            - INVALID_CREDENTIALS (401): email o contraseña incorrectos en el login, o token ausente, caducado o no válido.

            - SESSION_EXPIRED (401): POST /api/auth/refresh no puede renovar la sesión (refresh token ausente, caducado, revocado
            o reutilizado); hay que volver a iniciar sesión.

            - ACCESS_DENIED (403): autenticado, pero sin permisos (por ejemplo, no es ADMIN).

            - METHOD_NOT_ALLOWED (405): el método HTTP no está soportado por la ruta.

            - UNSUPPORTED_MEDIA_TYPE (415): el tipo de contenido no está soportado.

            - NOT_ACCEPTABLE (406): la cabecera Accept pide un formato que la API no produce; la API solo responde en JSON
            (este error también va en JSON).

            - RATE_LIMIT_EXCEEDED (429): demasiadas peticiones de login, registro o renovación de sesión (POST /api/auth/refresh)
            desde la misma IP, o demasiados intentos con la contraseña actual incorrecta en PUT /api/users/me/password (por
            cuenta: 5 cada 15 minutos); ver la cabecera Retry-After.

            - ACCOUNT_LOCKED (429): la cuenta está bloqueada temporalmente por logins fallidos; ver la cabecera Retry-After.

            - INTERNAL_ERROR (500 u otro 5xx): error no controlado en el servidor.'
          enum:
          - RESOURCE_NOT_FOUND
          - USER_ALREADY_EXISTS
          - MOVIE_ALREADY_IN_FAVORITES
          - MOVIE_NOT_IN_FAVORITES
          - VALIDATION_ERROR
          - INVALID_CREDENTIALS
          - DATA_INTEGRITY_VIOLATION
          - INTERNAL_ERROR
          - AMBIGUOUS_TITLE
          - ACCESS_DENIED
          - MALFORMED_REQUEST
          - METHOD_NOT_ALLOWED
          - UNSUPPORTED_MEDIA_TYPE
          - NOT_ACCEPTABLE
          - RATE_LIMIT_EXCEEDED
          - ACCOUNT_LOCKED
          - GENRE_ALREADY_EXISTS
          - GENRE_IN_USE
          - EPISODE_ALREADY_EXISTS
          - SERIES_ALREADY_IN_FAVORITES
          - SERIES_NOT_IN_FAVORITES
          - CSRF_REJECTED
          - SESSION_EXPIRED
          - CURRENT_PASSWORD_INCORRECT
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          description: 'Mensaje legible en español, pensado para mostrarlo al usuario. Puede cambiar: no se debe usar para
            tomar decisiones.'
          example: Película no encontrada
        path:
          type: string
          description: Ruta de la petición que produjo el error (sin los parámetros de consulta).
          example: /api/movies/42
        validationErrors:
          type: object
          additionalProperties:
            type: string
          description: 'Solo en los 400 con código VALIDATION_ERROR o CURRENT_PASSWORD_INCORRECT: un mensaje por cada campo
            o parámetro no válido (nombre del campo → mensaje). En el resto de errores no aparece (no se envía con null).'
          example:
            title: El título es obligatorio
        remainingAttempts:
          type: integer
          format: int32
          description: 'Solo en el 401 INVALID_CREDENTIALS del login: intentos que le quedan a la cuenta, con este fallo ya
            descontado. Nunca vale 0, porque el fallo que agota los intentos ya responde 429 ACCOUNT_LOCKED. En el resto de
            errores no aparece (no se envía con null).'
          example: 4
          minimum: 1
      required:
      - code
      - error
      - message
      - path
      - status
      - timestamp
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: 'Autenticación principal (navegador): el JWT de acceso viaja en la cookie HttpOnly streambox_token que
        fija POST /api/auth/login (Path=/api, 15 minutos; se renueva con POST /api/auth/refresh). Las peticiones no seguras
        (POST, PUT, PATCH, DELETE) autenticadas por esta cookie deben llevar además la cabecera X-Requested-With: StreamBox
        (defensa CSRF); si falta, 403 CSRF_REJECTED. Si la petición trae Authorization: Bearer, manda el Bearer y la cookie
        no se mira. Swagger UI no puede fijar esta cookie: para probar desde aquí usa bearerAuth.'
      name: streambox_token
      in: cookie
    bearerAuth:
      type: http
      description: 'Alternativa para clientes de API sin navegador: el mismo JWT de acceso (el valor de la cookie streambox_token)
        en la cabecera Authorization: Bearer. Con Bearer no se exige X-Requested-With (no hay riesgo de CSRF).'
      scheme: bearer
      bearerFormat: JWT